v25: remove OpenTracing support
