+++ go/flags/endtoend/vtclient.txt
@@ -21,9 +21,6 @@ Flags:
       --config-persistence-min-interval duration                    minimum interval between persisting dynamic config changes back to disk (if no change has occurred, nothing is done). (default 1s)
       --config-type string                                          Config file type (omit to infer config type from file extension).
       --count int                                                   DMLs only: Number of times each thread executes the query. Useful for simple, sustained load testing. (default 1)
-      --datadog-agent-host string                                   host to send spans to. if empty, no tracing will be done
-      --datadog-agent-port string                                   port to send spans to. if empty, no tracing will be done
-      --datadog-trace-debug-mode                                    enable debug mode for datadog tracing
       --grpc-auth-static-client-creds string                        When using grpc_static_auth in the server, this file provides the credentials to use to authenticate with server.
       --grpc-compression string                                     Which protocol to use for compressing gRPC. Default: nothing. Supported: snappy
       --grpc-dial-concurrency-limit int                             Maximum concurrency of grpc dial operations. This should be less than the golang max thread limit of 10000. (default 1024)
@@ -35,7 +32,6 @@ Flags:
       --grpc-max-message-size int                                   Maximum allowed RPC message size. Larger messages will be rejected by gRPC with the error 'exceeding the max size'. (default 16777216)
       --grpc-prometheus                                             Enable gRPC monitoring with Prometheus.
   -h, --help                                                        help for vtclient
-      --jaeger-agent-host string                                    host and port to send spans to. if empty, no tracing will be done
       --json                                                        Output JSON instead of human-readable table
       --keep-logs duration                                          keep logs for this long (using ctime) (zero to keep forever)
       --keep-logs-by-mtime duration                                 keep logs for this long (using mtime) (zero to keep forever)
@@ -62,8 +58,7 @@ Flags:
       --timeout duration                                            timeout for queries (default 30s)
       --tracer string                                               tracing service to use (default "noop")
       --tracing-enable-logging                                      whether to enable logging in the tracing service
-      --tracing-sampling-rate float                                 sampling parameter for traces; for jaeger this is passed as the sampler parameter (see --tracing-sampling-type), for opentelemetry/datadog it is a probability between 0.0 and 1.0 (default 0.1)
-      --tracing-sampling-type string                                sampling strategy to use for jaeger. possible values are 'const', 'probabilistic', 'rateLimiting', or 'remote' (default "const")
+      --tracing-sampling-rate float                                 sampling rate for traces as a probability between 0.0 and 1.0 (default 0.1)
       --use-random-sequence                                         use random sequence for generating [min-sequence-id, max-sequence-id)
       --v Level                                                     log level for V logs
   -v, --version                                                     print binary version
