--- go/trace/plugin_jaeger_test.go
@@ -1,35 +1,0 @@
-/*
-Copyright 2024 The Vitess Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package trace
-
-import (
-	"testing"
-
-	"github.com/stretchr/testify/require"
-)
-
-func TestNewJaegerTracerFromEnv(t *testing.T) {
-	tracingSvc, closer, err := newJagerTracerFromEnv("noop")
-	require.NoError(t, err)
-	require.NotEmpty(t, tracingSvc)
-	require.NotEmpty(t, closer)
-
-	tracingSvc, closer, err = newJagerTracerFromEnv("")
-	require.ErrorContains(t, err, "no service name provided")
-	require.Empty(t, tracingSvc)
-	require.Empty(t, closer)
-}
